+++ bill/discounts.go
@@ -190,7 +190,7 @@ func calculateDiscounts(lines []*Discount, cur currency.Code, sum num.Amount, rr
 			continue
 		}
 		l.Index = i + 1
-		if l.Percent != nil && !l.Percent.IsZero() {
+		if l.Percent != nil {
 			base := sum
 			if l.Base != nil {
 				base = l.Base.RescaleUp(zero.Exp() + linePrecisionExtra)